backend/requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
prisma==0.11.0
asyncpg==0.29.0

# Auth & Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
supabase==2.2.0

# Redis
redis==5.0.1

# AI & Embeddings
openai==1.3.5
tiktoken==0.5.1
langchain==0.0.340

# Utils
python-dotenv==1.0.0
httpx==0.25.2
tenacity==8.2.3

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
